Add GetRentalCommandHandler and refactor notification system.
New features:
- Implemented GetRentalCommandHandler to return rental info by ID. Customers can access only their rentals; managers can access all.
- Added a method in RentalNotificationTemplates to generate rental messages based on user role (isCustomer flag).

Notification system refactor:
- Simplified NotificationSender interface to a single sendNotification() method for flexibility and scalability.
- Added RentalNotificationListener and PaymentNotificationListener to handle notifications via NotificationSender implementations using event listeners.
- Introduced NotificationType enum to define all notification cases (e.g., NEW_RENTAL, RETURN_RENTAL, SUCCESSFUL_PAYMENT).
- Added RentalNotificationSender and PaymentNotificationSender, which handle specific notification types based on the NotificationType and related entity (Rental or Payment).
- Introduced TelegramBotService.notifyRecipients() method to send messages to both customers and managers, reducing code duplication.

Message creation changes:
- Added NotificationRecipient enum for all user roles receiving notifications (CUSTOMER, MANAGER).
- All notification messages (excluding command responses) are now returned as EnumMap<NotificationRecipient, String>, providing messages for each role.
