+++ src/main/java/com/carshare/rentalsystem/dto/payment/even/dto/PaymentCancelEventDto.java
@@ -1,0 +1,6 @@
+package com.carshare.rentalsystem.dto.payment.even.dto;
+
+import com.carshare.rentalsystem.model.Payment;
+
+public record PaymentCancelEventDto(Payment payment, Long userId) {
+}
