+++ src/main/java/com/carshare/rentalsystem/dto/payment/even/dto/PaymentSuccessEventDto.java
@@ -1,0 +1,6 @@
+package com.carshare.rentalsystem.dto.payment.even.dto;
+
+import com.carshare.rentalsystem.model.Payment;
+
+public record PaymentSuccessEventDto(Payment payment, Long userId) {
+}
