+++ src/main/java/com/carshare/rentalsystem/dto/rental/event/dto/RentalCreatedEventDto.java
@@ -1,0 +1,6 @@
+package com.carshare.rentalsystem.dto.rental.event.dto;
+
+import com.carshare.rentalsystem.model.Rental;
+
+public record RentalCreatedEventDto(Rental rental, Long userId) {
+}
