+++ src/main/java/com/carshare/rentalsystem/dto/rental/event/dto/RentalReturnEventDto.java
@@ -1,0 +1,6 @@
+package com.carshare.rentalsystem.dto.rental.event.dto;
+
+import com.carshare.rentalsystem.model.Rental;
+
+public record RentalReturnEventDto(Rental rental, Long userId) {
+}
