--- src/main/java/com/carshare/rentalsystem/notifications/NotificationType.java
@@ -1,6 +1,0 @@
-package com.carshare.rentalsystem.notifications;
-
-public enum NotificationType {
-    MANAGER,
-    CUSTOMER
-}
