--- src/main/java/com/carshare/rentalsystem/notifications/telegram/command/handler/TelegramCommandHandler.java
@@ -1,10 +1,0 @@
-package com.carshare.rentalsystem.notifications.telegram.command.handler;
-
-import com.pengrad.telegrambot.TelegramBot;
-import com.pengrad.telegrambot.model.Message;
-
-public interface TelegramCommandHandler {
-    String getCommand();
-
-    void handle(TelegramBot bot, Message message);
-}
