+++ src/main/java/com/carshare/rentalsystem/service/notifications/NotificationSender.java
@@ -1,0 +1,5 @@
+package com.carshare.rentalsystem.service.notifications;
+
+public interface NotificationSender {
+    void sendNotification(NotificationType type, Object notificationData, Long userId);
+}
