+++ src/main/java/com/carshare/rentalsystem/service/notifications/NotificationType.java
@@ -1,0 +1,9 @@
+package com.carshare.rentalsystem.service.notifications;
+
+public enum NotificationType {
+    NEW_RENTAL,
+    RETURN_RENTAL,
+    SUCCESSFUL_PAYMENT,
+    PAYMENT_CANCEL,
+    RENEW_PAYMENT
+}
