+++ src/main/java/com/carshare/rentalsystem/service/notifications/telegram/templates/NotificationRecipient.java
@@ -1,0 +1,6 @@
+package com.carshare.rentalsystem.service.notifications.telegram.templates;
+
+public enum NotificationRecipient {
+    CUSTOMER,
+    MANAGER
+}
